refactor: 베럴 패턴 적용으로 인한 변경사항
