+++ apps/web/src/entities/user/api/index.ts
@@ -1,0 +1,2 @@
+export { getUserInfo } from './get-user-info';
+export { getUserStatus } from './get-user-status';
