--- apps/web/src/entities/user/index.ts
@@ -1,6 +1,0 @@
-export { getUserStatus } from './api/get-user-status';
-export { END_POINT } from './model/end_point';
-export type { GetUserStatusResponse } from './model/types';
-export { USER_STATUS_QUERY_OPTIONS } from './queries/queries';
-export { USER_STATUS_QUERY_KEY } from './queries/query-key';
-export { default as VisaStatusCard } from './ui/visa-status-card/visa-status-card';
