+++ apps/web/src/entities/user/model/index.ts
@@ -1,0 +1,3 @@
+export { END_POINT } from './end_point';
+export type { GetUserStatusResponse } from './types';
+export type { GetUserInfoResponse } from './types';
