+++ apps/web/src/entities/user/queries/index.ts
@@ -1,0 +1,2 @@
+export { USER_QUERY_OPTIONS } from './queries';
+export { USER_QUERY_KEY } from './query-key';
