+++ apps/web/src/entities/user/queries/queries.ts
@@ -1,10 +1,7 @@
 import { queryOptions } from '@tanstack/react-query';
 
-import { getUserStatus } from '@entities/user/api/get-user-status';
-
-import { getUserInfo } from '../api/get-user-info';
-import { USER_QUERY_KEY } from './query-key';
-import { USER_STATUS_QUERY_KEY } from './query-key';
+import { getUserInfo, getUserStatus } from '@entities/user/api';
+import { USER_QUERY_KEY } from '@entities/user/queries';
 
 export const USER_QUERY_OPTIONS = {
   GET_USER_INFO: () => {
@@ -13,12 +10,9 @@ export const USER_QUERY_OPTIONS = {
       queryFn: getUserInfo,
     });
   },
-};
-
-export const USER_STATUS_QUERY_OPTIONS = {
   GET_USER_STATUS: () => {
     return queryOptions({
-      queryKey: USER_STATUS_QUERY_KEY.USER_STATUS(),
+      queryKey: USER_QUERY_KEY.USER_STATUS(),
       queryFn: getUserStatus,
     });
   },
