+++ apps/web/src/entities/user/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as VisaStatusCard } from './visa-status-card/visa-status-card';
